feat(evaluation):add utils for reusable functions
