+++ evaluation/schema/validate_qa_dataset.py
@@ -8,28 +8,10 @@ Purpose: Schema validation for QA datasets used for this project.
 import json
 import logging
 from typing import Any, Dict, List, Optional
+from evaluation.utils import load_json
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-
-def load_json(path: str) -> Optional[List[Dict[str, Any]]]:
-    """
-    Loads a JSON file and returns its content.
-
-    Args:
-        path (str): Path to the JSON file.
-    
-    Returns:
-        Optional[List[Dict[str, Any]]]: Parsed JSON data as list of dictionaries
-        or None if loading fails.
-    """
-    try:
-        with open(path, "r", encoding="utf-8") as f:
-            return json.load(f)
-    except Exception as e:
-        logger.error(f"❌ Failed to JSON from {path}: {e}")
-        return None
 
 
 def validate_qa_schema(
